tableai_plugins/pyproject.toml
[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
bump = true

[project]
name = "tableai-plugins"
dynamic = ["version"]
description = "PDF Table extraction UV monorepo: core lib + backend + optional plugins (CPU/GPU)."
authors = [
        { name = "John Schlafly", email = "john@schlafdata.com" },
]
readme = "README.md"
requires-python = ">=3.11,<4.0"

dependencies = [
    "pydantic==2.9.2",
    "torch",
    "torchvision",
    "huggingface-hub==0.34.4",
    "transformers>=4.41",
    "PyYAML==6.0.2",
]

[project.urls]
Homepage = "https://github.com/jschlafdata/tableai"
Source = "https://github.com/jschlafdata/tableai"

[project.optional-dependencies]
yolo = [
  "ultralytics @ git+https://github.com/THU-MIG/yolov10.git",
  "opencv-python>=4.10.0.84",
  "scipy>=1.14.0"
]

dit = [
  "timm==1.0.12",
  "accelerate>=0.29",
  "umap-learn==0.5.7",
  "tqdm==4.67.1",
]

all = [
  "ultralytics @ git+https://github.com/THU-MIG/yolov10.git",
  "opencv-python>=4.10.0.84",
  "scipy>=1.14.0",
  "timm==1.0.12",
  "accelerate>=0.29",
  "umap-learn==0.5.7",
  "tqdm==4.67.1",
]


[tool.uv]
cache-keys = [
    { file = "pyproject.toml" }, 
    { git = { commit = true, tags = true }},
    { file = "**/*.toml" }
]

[tool.hatch.build.targets.wheel]
packages = ["tableai_plugins"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]

[project.entry-points."tableai.plugins"]
yolo = "tableai_plugins.yolo:YOLOPlugin"
dit  = "tableai_plugins.dit:DiTPlugin"